fault_web/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
eventlet==0.36.1
pymongo==4.8.0
dnspython==2.6.1
python-dotenv==1.0.1
python-socketio==5.11.4